mpvq: fix queuing after deprecated input-file
mpv finally (to my frustration) removed the --input-file option. So I
had to rewrite everything to work with Unix domain sockets, which means
introducing a dependency on netcat with the -U flag (I think netcat-bsd
or something like that). I can't think of another way to do it at the
moment.


Former-commit-id: 3ebf5018b6824bf803816cb91caaa1a0de43fdb3